SCT-63: Review, extracting icons, making the app work because of a missing file
